♻️ refactor(sync): migrate contributors to descriptor catalog
