+++ sharedLogic/src/commonMain/kotlin/com/calypsan/listenup/client/data/sync/domains/SyncDomainCatalog.kt
@@ -39,6 +39,7 @@ internal fun syncDomainCatalog(
                 shelvesDomain(database = database),
                 playbackPositionsDomain(database = database),
                 seriesDomain(database = database),
+                contributorsDomain(database = database, imageStorage = imageStorage),
                 booksDomain(
                     database = database,
                     mapper = mapper,
