+++ sharedLogic/src/commonMain/kotlin/com/calypsan/listenup/client/data/sync/handlers/SyncHandlerSupport.kt
@@ -12,7 +12,7 @@ import kotlin.coroutines.cancellation.CancellationException
  * re-thrown — cancellation is not a sync failure.
  *
  * The canonical apply-wrapper for every [com.calypsan.listenup.client.data.sync.SyncDomainHandler]:
- * `ComposedSyncDomainHandler`, `ContributorSyncDomainHandler`
+ * `ComposedSyncDomainHandler` and the remaining hand-written handlers
  * all route their `onEvent` / `onCatchUpItem` work through it.
  */
 internal suspend fun TransactionRunner.applyEventAtomically(
