+++ sharedLogic/src/jvmTest/kotlin/com/calypsan/listenup/client/data/sync/ContributorDeleteCascadeE2ETest.kt
@@ -37,7 +37,7 @@ private const val BOOK_TWO_ID = "cascade-b2"
  * All three events must land in client Room: both books' `book_contributors`
  * junction rows are gone (the [BookMirrorApply.applyContributors] path
  * replaces the junction set on every book upsert), and the contributor row carries
- * `deletedAt != null` (the [ContributorSyncDomainHandler]'s `Deleted` branch sets
+ * `deletedAt != null` (the `contributors` domain's `Deleted` branch sets
  * the tombstone). The poll witness is the empty `getByBookId` join for both books
  * combined with a non-null `deletedAt` on the contributor row — the cascade is
  * only fully applied when all three conditions hold.
@@ -104,7 +104,7 @@ class ContributorDeleteCascadeE2ETest :
                 // row is tombstoned. The two `book.Updated` events drive the junction wipe via
                 // `BookMirrorApply.applyContributors` (which deleteContributorsForBook +
                 // re-inserts the empty list); the `contributor.Deleted` event drives the
-                // `deletedAt` write via `ContributorSyncDomainHandler.softDelete`.
+                // `deletedAt` write via the `contributors` domain's soft-delete.
                 withTimeout(ROUND_TRIP_TIMEOUT_SECONDS.seconds) {
                     while (true) {
                         val bookOneEmpty = clientDatabase.contributorDao().getByBookId(BOOK_ONE_ID).isEmpty()
