+++ sharedLogic/src/jvmTest/kotlin/com/calypsan/listenup/client/data/sync/testing/WithTagSyncEngineAgainstServer.kt
@@ -22,7 +22,7 @@ import com.calypsan.listenup.client.data.sync.domains.toHandler
 import com.calypsan.listenup.client.data.sync.domains.booksDomain
 import com.calypsan.listenup.client.data.sync.domains.librariesDomain
 import com.calypsan.listenup.client.data.sync.handlers.BookTagSyncDomainHandler
-import com.calypsan.listenup.client.data.sync.handlers.ContributorSyncDomainHandler
+import com.calypsan.listenup.client.data.sync.domains.contributorsDomain
 import com.calypsan.listenup.client.data.sync.domains.libraryFoldersDomain
 import com.calypsan.listenup.client.data.sync.handlers.ListeningEventSyncDomainHandler
 import com.calypsan.listenup.client.data.sync.domains.playbackPositionsDomain
@@ -191,10 +191,8 @@ internal fun withTagSyncEngineAgainstServer(block: suspend TagSyncEngineScope.()
                 mapper = BookEntityMapper(),
                 imageStorage = stubImageStorage(),
             ).toHandler(transactionRunner = RoomTransactionRunner(clientDb), registry = registry)
-            ContributorSyncDomainHandler(
-                database = clientDb,
+            contributorsDomain(database = clientDb, imageStorage = stubImageStorage()).toHandler(
                 transactionRunner = RoomTransactionRunner(clientDb),
-                imageStorage = stubImageStorage(),
                 registry = registry,
             )
             seriesDomain(database = clientDb).toHandler(
